+++ pylyskom/komsession.py
@@ -21,9 +21,12 @@ from .cachedconnection import Client, CachingPersonClient
 from .stats import stats
 
 from .datatypes import (
+    AuxItem,
     AuxItemInput,
     ConfType,
+    Conference,
     CookedMiscInfo,
+    ExtendedConfType,
     MIC_COMMENT,
     MIC_FOOTNOTE,
     MICommentTo,
@@ -31,14 +34,13 @@ from .datatypes import (
     MIR_CC,
     MIR_TO,
     MIRecipient,
+    Membership,
     MembershipType,
     PersonalFlags,
-    AuxItem,
     TextStat,
     UConference,
-    Conference,
-    Membership,
-    first_aux_items_with_tag)
+    first_aux_items_with_tag,
+)
 
 class KomSessionException(Exception): pass
 class KomSessionNotConnected(KomSessionException): pass
@@ -51,17 +53,32 @@ class NoRecipients(KomSessionError): pass
 MIRecipient_str_to_type = { 'to': MIR_TO,
                             'cc': MIR_CC,
                             'bcc': MIR_BCC }
-        
+
 MICommentTo_str_to_type = { 'comment': MIC_COMMENT,
                             'footnote': MIC_FOOTNOTE }
 
 
 
 
+# KomPerson is intended for the full person-stat with all its info, but is
+# not really used yet.
 class KomPerson:
     def __init__(self, pers_no, username: str):
         self.pers_no = pers_no
         self.username = username
+
+    def __repr__(self):
+        return f"KomPerson({self.pers_no!r}, {self.username!r})"
+
+# KomPersonName is intended for when only the pers-no and name are
+# wanted.
+class KomPersonName:
+    def __init__(self, pers_no, username: str):
+        self.pers_no = pers_no
+        self.username = username
+
+    def __repr__(self):
+        return f"KomPersonName({self.pers_no!r}, {self.username!r})"
 
 
 class KomMembershipUnread:
@@ -83,16 +100,37 @@ class KomAuxItem:
         self.creator = creator
 
 
+# KomConferenceName is intended for when only the conf-no and name are
+# wanted.
+class KomConferenceName:
+    def __init__(self, conf_no, name):
+        self.conf_no = conf_no
+        self.name = name
+
+
 # U stands for micro (as in the Protocol A spec)
 class KomUConference:
     def __init__(self, conf_no, *,
-                 uconf: UConference):
+                 uconf: Optional[UConference] = None,
+                 name: Optional[str] = None,
+                 type: Optional[ExtendedConfType] = None,
+                 highest_local_no: Optional[int] = None,
+                 nice: Optional[int] = None):
         self.conf_no = conf_no
 
-        self.name = uconf.name.decode('latin1')
-        self.type = uconf.type
-        self.highest_local_no = uconf.highest_local_no
-        self.nice = uconf.nice
+        if uconf is None:
+            self.name = name
+            self.type = type
+            self.highest_local_no = highest_local_no
+            self.nice = nice
+        else:
+            if isinstance(uconf.name, six.binary_type):
+                self.name = uconf.name.decode('latin1')
+            else:
+                self.name = uconf.name
+            self.type = uconf.type
+            self.highest_local_no = uconf.highest_local_no
+            self.nice = uconf.nice
 
 
 class KomConference:
@@ -452,6 +490,10 @@ class KomSession(object):
         return conf_no
 
     @check_connection
+    def delete_conference(self, conf_no):
+        self._client.request(requests.ReqDeleteConf(conf_no))
+
+    @check_connection
     def lookup_name(self, name, want_pers, want_confs):
         if isinstance(name, six.binary_type):
             name = name.decode('utf-8')
@@ -589,51 +631,6 @@ class KomSession(object):
             super_conf=super_conf,
             aux_items=aux_items)
 
-
-    # TODO: get_conference() can fail with the following message (in
-    # browser) if any of the related (like created-by) conferences is
-    # secret or deleted:
-    #
-    #   Failed to get conference.
-    #   {"error_code":9,"error_msg":"UndefinedConference","error_status":"547","error_type":"protocol-a"}
-    #
-    # The error above is from reading "Butiker erfarenhetsutbyte" (möte 604).
-    # The get-uconf-stat that fails is for the "created by" and "supermeeting" (möte 547), which
-    # has been deleted.
-    #
-    # get_conference() call will request generate get-uconf-stat calls
-    # for all related meetings and it's one of those that fails. We
-    # should handle that such meetings might be secret or deleted,
-    # without failing the get_conference() call.
-    #
-    # This is how the meeting looks like in the elisp client:
-    #
-    #   Status för möte Butiker erfarenhetsutbyte (604)
-    #
-    #   Skapat av person                       547 (Person 547 (finns inte).)
-    #   Skapad:                   1991-08-29 10:41
-    #   Antal medlemmar:                       419
-    #   Hemliga medlemmar:                       Hemliga medlemmar är inte tillåtna
-    #   Anonyma inlägg:                          Anonyma inlägg är inte tillåtna
-    #   Livslängd på inlägg:                    77 dagar
-    #   Minsta livslängd för kommenterade inlägg: 0 dagar
-    #   Lägsta existerande lokala nummer:       29
-    #   Högsta existerande lokala nummer:    38690
-    #   Tid för senaste inlägg:         idag 08:57 (står det i din cache)
-    #   Lapp på dörren i text nummer:            0
-    #   Supermöte:                             547 (Möte 547 (finns inte).)
-    #   Tillåtna författare:                     0 (Alla)
-    #   Organisatör:                          3343 (Magnus Bark (en rutten själ i en halvrund kropp))
-    #   Presentation:                     10785008
-    #   FAQ i inlägg:                      8359413 "Något om internationella postorderköp" [*]
-    #   FAQ i inlägg:                      8326719 "Skribofont" [*]
-    #   FAQ i inlägg:                     10176893 "sida typ bugsoft" [*]
-    #   FAQ i inlägg:                     10179028 "Hämta ut paket MED avi minst en dag tidigare" [*]
-    #
-    #
-    # See also:
-    # https://www.lysator.liu.se/lyskom/protocol/11.1/protocol-a.html#get-uconf-stat
-    #
     @check_connection
     def get_conference(self, conf_no, micro=True):
         conf_no = int(conf_no)